\name{bpBagging}
\alias{bpBagging}
\title{Build model and predict results using bagging for classification}
\usage{
bpBagging(data, ...)
}
\arguments{
  \item{data}{output of readData or readData2.}

  \item{...}{further arguments to be passed to bagging}
}
\description{
Build model and predict results using bagging for
classification
}
\examples{
# bpBagging(data=mydata)
}
\seealso{
\code{\link{readData}}

\code{\link{readData2}}

\code{\link{writeData}}

\code{\link{plotData}}
}

