\name{bpCtree}
\alias{bpCtree}
\title{Build model and predict results using conditional inference trees}
\usage{
bpCtree(data, ...)
}
\arguments{
  \item{data}{output of readData or readData2.}

  \item{...}{further arguments to be passed to ctree}
}
\description{
Build model and predict results using conditional inference
trees
}
\examples{
# bpCtree(data=mydata)
}
\seealso{
\code{\link{readData}}

\code{\link{readData2}}

\code{\link{writeData}}

\code{\link{plotData}}
}

