\name{bpLogisticRegression}
\alias{bpLogisticRegression}
\title{Build model and predict results using logistic regression}
\usage{
bpLogisticRegression(data, ...)
}
\arguments{
  \item{data}{output of readData or readData2.}

  \item{...}{further arguments to be passed to glm}
}
\description{
Build model and predict results using logistic regression
}
\examples{
# bpLogisticRegression(data=mydata)
}
\seealso{
\code{\link{readData}}

\code{\link{readData2}}

\code{\link{writeData}}

\code{\link{plotData}}
}

