\name{bpRandomForest}
\alias{bpRandomForest}
\title{Build model and predict results using random forest}
\usage{
bpRandomForest(data, ...)
}
\arguments{
  \item{data}{output of readData or readData2.}

  \item{...}{further arguments to be passed to
  randomForest}
}
\description{
Build model and predict results using random forest
}
\examples{
# bpRandomForest(data=mydata)
}
\seealso{
\code{\link{readData}}

\code{\link{readData2}}

\code{\link{writeData}}

\code{\link{plotData}}
}

