% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readPedMap}
\alias{readPedMap}
\title{Reads PED/MAP file in PLINK format}
\usage{
readPedMap(ped, map, ped2 = NULL, map2 = NULL, prop = 0.8, ...)
}
\arguments{
\item{ped}{ped file contains training data or both training and testing data}

\item{map}{map file matches the ped file}

\item{ped2}{ped file contains testing data}

\item{map2}{map file matches the ped2 file}

\item{prop}{proportion of the data will be used as training data}

\item{...}{further arguments to be passed to read.table or read.table}
}
\description{
Reads PED/MAP file in PLINK format
}
\examples{
# readPedMap(ped="all.ped",map="all.map",prop=0.7)
# readPedMap(ped="train.ped",map="train.map",ped2="test.ped",map2="test.map")
}

